+++ .storybook/preview.js
@@ -7,3 +7,9 @@ import { withHTML } from "@whitespace/storybook-addon-html/html"
 
 addDecorator(withA11y)
 addDecorator(withHTML)
+
+import { initToggle } from "../src/js/libs/Toggle"
+
+setTimeout(() => {
+  initToggle()
+}, 2000)