+++ src/02-molecules/teaser/stories/teaser.data.js
@@ -1,12 +1,6 @@
+import { imageData } from "../../../01-atoms/image/stories/image.data"
+
 export const teaserData = {
-  img: {
-    dataSrcset: `
-      https://placehold.it/358x260 1@,
-      https://placehold.it/716x520 2@,
-      https://placehold.it/1074x780 3@,
-    `,
-    dataSrc: 'https://placehold.it/358x260',
-    src: 'https://placehold.it/358x260',
-  },
-  title: 'Küchenabdeckungen'
-};
+  img: imageData,
+  title: "Küchenabdeckungen"
+}