+++ src/03-organisms/overview-teasers/stories/overview-teasers.data.js
@@ -1,180 +1,86 @@
+import { imageData } from "../../../01-atoms/image/stories/image.data"
+
 export const teaserData = [
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Küchenabdeckungen"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Bodenbeläge und Treppen (Innen)"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Bodenbeläge und Treppen (Aussen)"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Elementbau (Innen)"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Elementbau (Aussen)"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Fassaden"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Gärten"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Grabsteine"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Berner Sandstein"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Oberflächenbearbeitung"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Badezimmer"
   },
   {
     img: {
-      dataSrcset: `
-      https://placehold.it/358x260 358w,
-      https://placehold.it/716x520 716w,
-      https://placehold.it/1074x780 1074w,
-      https://placehold.it/1432x1040 1432w,
-      https://placehold.it/1790x1300 1790w,
-    `,
-      dataSrc: "https://placehold.it/358x260",
-      src: "https://placehold.it/358x260",
+      ...imageData,
       sizes: "(min-width: 720px) 30vw, 100vw"
     },
     title: "Wasseraufbereitung & Abwassertechnik"