+++ src/03-organisms/teasers-grid/index.js
@@ -1,17 +1,19 @@
 import React from "react"
 import Teaser from "../../02-molecules/teaser"
 
-const OverviewTeasers = ({ teasers = [] }) => (
-  <ul className="o-overview-teasers">
+const TeasersGrid = ({ teasers = [], teaserClasses }) => (
+  <ul className="o-teasers-grid">
     {teasers &&
       teasers.map(teaser => (
         <Teaser
           key={encodeURIComponent(teaser.title)}
           imageData={teaser.img}
           title={teaser.title}
+          text={teaser.text}
+          classes={teaserClasses}
         />
       ))}
   </ul>
 )
 
-export default OverviewTeasers
+export default TeasersGrid