+++ src/main.js
@@ -1,14 +1,16 @@
 // import React from 'react';
 
-(() => {
+;(() => {
   // alert('hi there');
   // const btn = document.querySelector('a-btn');
   // btn.addEventListener('click', alter('hi there'));
 
   setTimeout(() => {
-    const btn = document.querySelector('a-btn');
-    if (btn) {
-      btn.addEventListener('click', () => alert('hi there'));
-    }
-  }, 2000);
-})();
+    // const btn = document.querySelector('a-btn');
+    // if (btn) {
+    //   btn.addEventListener('click', () => alert('hi there'));
+    // }
+
+    console.log("Hello From main.js")
+  }, 2000)
+})()